Implement Netlink
With this commit i'm introducing the netlink protocol to read routes
from the linux kernel routing table to the locRIB.
This also adds the functionality to insert new routes to the linux
kernel routing stack which where learned by other protocol (like e.g.
BGP).

This is done by the protocols/netlink/netlink.go object.
The netlink.go file acts as an facade around the netlink_reader.go and
netlink_writer.go.

netlink_reader itself reads in a defined reding interval new routes from
the kernel and propagates new routes to all subscribed clients.
The netlink reader can also be filtered using the same filter-api as in
adjRibIn/adjRibOut for BGP.
To read new routes to the locRIB you have to register the locRIB to the
netlink_reader using the ClientManager interface, just like you would
register the locRIB to the adjRibIn.

netlink_writer must be subscribed to the locRIB in order to write the
routes from locRIB into the kernel.
This useses the ClientManager interface, just like you would register
the adjRibOut to the locRIB.
netlink_writer itself currently does not support filters, since all
routes form the locRIB should be written to the kernel.
The filter functionality could be easily added here as well.

The netlink_writer currently does not support ecmp. This is subject of
an other commit.

The current implementation could be considered as work in progress, but
the code is mature enough to start a review on it.

+++ net/ip_test.go
@@ -13,7 +13,7 @@ func TestCompare(t *testing.T) {
 		name     string
 		ip       IP
 		other    IP
-		expected int
+		expected int8
 	}{
 		{
 			name: "equal",
